requirements.txt
psutil
customtkinter
kafka-python
